# add '_thisArgs', '_thisId', '_thisType' and '_thisFnc'  and  to CBA events