--- a/addons/events/CfgFunctions.hpp
+++ b/addons/events/CfgFunctions.hpp
@@ -46,6 +46,10 @@ class CfgFunctions {
                 description = "Registers an event handler for a specific CBA event.";
                 file = "\x\cba\addons\events\fnc_addEventHandler.sqf";
             };
+            class addEventHandlerArgs {
+                description = "Registers an event handler for a specific CBA event with arguments.";
+                file = "\x\cba\addons\events\fnc_addEventHandlerArgs.sqf";
+            };
             class removeEventHandler {
                 description = "Removes an event handler previously registered with CBA_fnc_addEventHandler.";
                 file = "\x\cba\addons\events\fnc_removeEventHandler.sqf";