Closes: #374

Adds the same meta data `CBA_fnc_addBISEventHandler` has to the CBA events.

- adds optional parameter to `CBA_fnc_addEventHandler` to assign arguments to the event.

`_thisArgs` - Arguments passed when adding the event handler.
`_thisId` - Id that can be used to remove the event via `CBA_fnc_removeEventHandler`
`_thisType` - the name of the event ("testEvent" in `["testEvent", ...] call CBA_fnc_addEventHandler`)
`_thisFnc` - function that is executed when the event happens

Note: If the arguments passed to `CBA_fnc_addEventHandler` are an array, they are stored as array reference and not as copy. That means you can modify these arguments even after the event was added.

~~Needs some more testing, will do later ...~~