@commy2 2016-06-11
One time event handler printing "hello world"

```
["test1", {
    systemChat str _thisArgs;
    [_thisType, _thisId] call CBA_fnc_removeEventHandler
}, "hello world"] call CBA_fnc_addEventHandler;

...

"test1" call CBA_fnc_localEvent
```

Problem is that you are modifing an array GVAR(eventsNamespace) while iterating through it via `forEach` which is undefined ... Searching for a work-around.