@PabstMirror 2016-06-11
I'm really not sure if any of this is necessary, my main concern is that there will be a performance cost to adding these args.

For 10x localEvent calls for something with 3 added ehs:

Old: 0.218436 ms 0.216403 ms  0.217628 ms
New: 0.326691 ms 0.327654 ms 0.324465 ms