@commy2 2016-06-11
Good point. It's performance vs. functionality. I have no idea with what we should go here.
+0.10 ms isn't that bad though honestly