@PabstMirror 2016-06-11
ACE calls multiple EHs for each bullet fired, so 50% increase isn't good

I think we can do most of what we need by using a function wrapper. Check out my demo code here (not tested, more of an idea): https://github.com/CBATeam/CBA_A3/compare/addEventHandlerWrapper?expand=1