@commy2 2016-06-11
Is it a 50% increase even when measuring with the actual event function you are calling?