@PabstMirror 2016-06-11
```
Prep:
["test", {x1=_this;}] call CBA_fnc_addEventHandler;
["test", {x2=_this;}] call CBA_fnc_addEventHandler;
["test", {x3=_this;}] call CBA_fnc_addEventHandler;

Test:
["test", []] call CBA_fnc_localEvent;
["test", []] call CBA_fnc_localEvent;
["test", []] call CBA_fnc_localEvent;
["test", []] call CBA_fnc_localEvent;
["test", []] call CBA_fnc_localEvent;
["test", []] call CBA_fnc_localEvent;
["test", []] call CBA_fnc_localEvent;
["test", []] call CBA_fnc_localEvent;
["test", []] call CBA_fnc_localEvent;
["test", []] call CBA_fnc_localEvent;
```