@commy2 2016-06-11
My point was that `{x1=_this;}` is a extremly simple function. The bigger the event function is, the smaller the share of the overhead of `CBA_fnc_localEvent` becomes.