@commy2 2016-06-11
@PabstMirror I read the link you posted.
I like that idea. It would mean having no additional overhead by default, but an alternative addEventHandler function that would apply all these variables.
Any naming suggestions for `FUNC(addBigEH)`?

I'd still have to copy the array though, but that is only once per event call and not once per event per call.