@Killswitch00 2016-06-12
`addEventHandlerArgs`, `addArgsEventHandler`

An alternative/wrapper function sounds like the best way forward, so as to not compromise performance of the original function. Now, to think of a suitable name...