@commy2 2016-06-12
I'm going with `addEventHandlerArgs` instead of the usual `addXEventHandler`, because it isn't actually a new thing on it's own. It also uses the same `removeEventHandler` function.